requirements.txt
asgiref==3.5.2
boto3==1.26.48
botocore==1.29.48
coverage==7.0.5
dj-database-url==0.5.0
Django==3.2
django-allauth==0.41.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-storages==1.13.2
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.5.0
gunicorn==20.1.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
oauthlib==3.2.2
Pillow==9.3.0
pluggy==1.0.0
psycopg2==2.9.5
pytest==7.2.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python3-openid==3.2.0
pytz==2022.6
requests-oauthlib==1.3.1
s3transfer==0.6.0
sqlparse==0.4.3
stripe==5.0.0
